refactor: 불필요한 주석, 메서드 삭제, 띄어쓰기 컨벤션
